post refactoring cleanup
