+++ packages/cli/src/lib/discover-files.ts
