+++ packages/cli/src/lib/get-copy-static-plan.ts
