+++ packages/cli/src/lib/render-atom.ts
@@ -1,9 +1,9 @@
 import { Feed } from "feed";
 import { cliVersion } from "../utils/version";
-import { CACHE_FILENAME } from "./cache";
 import type { Config } from "./get-config";
 import type { EnrichedArticle, EnrichedSource } from "./enrich";
 import { getHostnameFromUrl } from "../utils/get-hostname-from-url";
+import { CACHE_FILENAME } from "./path-constants";
 
 export const FEED_FILENAME = "feed.atom";
 export const INDEX_FILENAME = "index.html";
